requirements.txt
Django==2.2.15
django-cors-headers==3.5.0
django-environ==0.4.5
gunicorn==20.0.4
psycopg2==2.8.5
django-filter==2.4.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
model-mommy==2.0.0
pydot==1.4.1
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
